update frr: remove old files
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@65906 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/frr/tests/topotests/mgmt_notif/test_ds_notify.py
@@ -12,19 +12,30 @@ import json
 import logging
 import os
 import re
-import time
 
 import pytest
+from lib.common_config import step
 from lib.topogen import Topogen
 from lib.topotest import json_cmp
 from munet.base import Timeout
-from munet.testing.util import readline, waitline
+from munet.testing.util import readline, retry, waitline
 from oper import check_kernel_32
 
-pytestmark = [pytest.mark.ripd, pytest.mark.staticd, pytest.mark.mgmtd]
+from lib import topotest
+
+pytestmark = [pytest.mark.staticd, pytest.mark.mgmtd]
 
 CWD = os.path.dirname(os.path.realpath(__file__))
 FE_CLIENT = CWD + "/../lib/fe_client.py"
+BE_CLIENT = "/usr/lib/frr/mgmtd_testc"
+
+
+def have_beclient(r1):
+    if hasattr(have_beclient, "has"):
+        return have_beclient.has
+    rc, _, _ = r1.cmd_status(BE_CLIENT + " --help")
+    have_beclient.has = True if not rc else False
+    return have_beclient.has
 
 
 @pytest.fixture(scope="module")
@@ -32,7 +43,7 @@ def tgen(request):
     "Setup/Teardown the environment and provide tgen argument to tests"
 
     topodef = {
-        "s1": ("r1", "r2"),
+        "s1": ("r1",),
     }
 
     tgen = Topogen(topodef, request.module.__name__)
@@ -40,7 +51,7 @@ def tgen(request):
 
     router_list = tgen.routers()
     for _, router in router_list.items():
-        router.load_frr_config("frr.conf")
+        router.load_frr_config("frr-ds.conf")
 
     tgen.start_router()
     yield tgen
@@ -75,7 +86,7 @@ def get_op_and_json(output):
 
 # Wait for specific OP, path and maybe json, path must match exactly,
 # return the path and json data (or None for DELETE)
-def wait_op_json(f, op, path, json_match=None, exact=False, timeout=10):
+def wait_op_json(f, op, path, json_match=None, exact=False, timeout=30):
     to = Timeout(timeout)
     jexp = json.loads(json_match) if isinstance(json_match, str) else json_match
     while not to.is_expired():
@@ -99,6 +110,14 @@ def wait_op_json(f, op, path, json_match=None, exact=False, timeout=10):
         logging.debug("no json match: %s: continue", jo)
 
 
+@retry(retry_timeout=30, initial_wait=0.5)
+def check_backend_xpath_registry(r1, repath):
+    output = r1.cmd_raises('vtysh -c "show mgmt backend-yang-xpath-registry notify"')
+    if re.match("notify: ?" + repath, output):
+        return "missing notify registration for " + repath
+    return None
+
+
 def test_frontend_datastore_notification(tgen):
     if tgen.routers_have_failure():
         pytest.skip(tgen.errors)
@@ -111,32 +130,25 @@ def test_frontend_datastore_notification(tgen):
     p = r1.popen(
         [
             FE_CLIENT,
+            "--verbose",
+            "--log=fe_client.log",
             "--datastore",
-            "--notify-count=2",
+            "--notify-count=0",
             "--listen=/frr-interface:lib/interface/state",
         ]
     )
-    assert waitline(p.stderr, "Connected", timeout=10)
-
-    r1.cmd_raises("ip link set r1-eth0 mtu 1200")
-
-    # {"frr-interface:lib":{"interface":[{"name":"r1-eth0","state":{"if-index":2,"mtu":1200,"mtu6":1200,"speed":10000,"metric":0,"phy-address":"ba:fd:de:b5:8b:90"}}]}}
 
     try:
-        # Wait for FE client to exit
-        output, error = p.communicate(timeout=10)
-        notifs = get_op_and_json(output)
-        op, path, data = notifs[1]
+        wait_op_json(p.stdout, "SYNC", "/frr-interface:lib/interface/state", None)
 
-        assert op == "REPLACE"
-        assert path.startswith("/frr-interface:lib/interface[name='r1-eth0']/state")
-
-        jsout = json.loads(data)
-        expected = json.loads(
-            '{"frr-interface:lib":{"interface":[{"name":"r1-eth0","state":{"mtu":1200}}]}}'
+        r1.cmd_raises("ip link set r1-eth0 mtu 1200")
+        wait_op_json(
+            p.stdout,
+            "REPLACE",
+            "/frr-interface:lib/interface[name='r1-eth0']/state/mtu",
+            '{"frr-interface:lib":{"interface":[{"name":"r1-eth0","state":{"mtu":1200}}]}}',
         )
-        result = json_cmp(jsout, expected)
-        assert result is None
+
     finally:
         p.kill()
         r1.cmd_raises("ip link set r1-eth0 mtu 1500")
@@ -150,41 +162,31 @@ def test_backend_datastore_update(tgen):
 
     check_kernel_32(r1, "11.11.11.11", 1, "")
 
-    be_client_path = "/usr/lib/frr/mgmtd_testc"
-    rc, _, _ = r1.cmd_status(be_client_path + " --help")
-
-    if rc:
+    if not have_beclient(r1):
         pytest.skip("No mgmtd_testc")
-
-    # Watch the mgmtd log for the BE subscribing
-    mlogp = r1.popen(["/usr/bin/tail", "-n0", "-f", f"{r1.rundir}/mgmtd.log"])
 
     # Start our BE client in the background
     p = r1.popen(
         [
-            be_client_path,
-            "--timeout=20",
-            "--log=file:/dev/stderr",
+            BE_CLIENT,
+            "--log=file:mgmtd-testc.log",
+            "--notify-count=0",
             "--datastore",
             "--listen",
             "/frr-interface:lib/interface",
         ]
     )
-    assert waitline(mlogp.stdout, 'now known as "mgmtd-testc"', timeout=10)
-    mlogp.kill()
 
-    r1.cmd_raises("ip link set r1-eth0 mtu 1200")
     try:
-        expected = json.loads(
-            '{"frr-interface:lib":{"interface":[{"name":"r1-eth0","state":{"mtu":1200}}]}}'
-        )
+        wait_op_json(p.stdout, "SYNC", "/frr-interface:lib/interface", None)
 
-        output, error = p.communicate(timeout=10)
-        notifs = get_op_and_json(output)
-        op, path, data = notifs[0]
-        jsout = json.loads(data)
-        result = json_cmp(jsout, expected)
-        assert result is None
+        r1.cmd_raises("ip link set r1-eth0 mtu 1200")
+        wait_op_json(
+            p.stdout,
+            "REPLACE",
+            "/frr-interface:lib/interface[name='r1-eth0']/state/mtu",
+            '{"frr-interface:lib":{"interface":[{"name":"r1-eth0","state":{"mtu":1200}}]}}',
+        )
     finally:
         p.kill()
         r1.cmd_raises("ip link set r1-eth0 mtu 1500")
@@ -198,21 +200,14 @@ def test_backend_datastore_add_delete(tgen):
 
     check_kernel_32(r1, "11.11.11.11", 1, "")
 
-    be_client_path = "/usr/lib/frr/mgmtd_testc"
-    rc, _, _ = r1.cmd_status(be_client_path + " --help")
-
-    if rc:
+    if not have_beclient(r1):
         pytest.skip("No mgmtd_testc")
-
-    # Watch the mgmtd log for the BE subscribing
-    mlogp = r1.popen(["/usr/bin/tail", "-n0", "-f", f"{r1.rundir}/mgmtd.log"])
 
     # Start our BE client in the background
     p = r1.popen(
         [
-            be_client_path,
-            "--timeout=60",
-            "--log=file:/dev/stderr",
+            BE_CLIENT,
+            "--log=file:mgmtd-testc.log",
             "--notify-count=0",
             "--datastore",
             "--listen",
@@ -220,22 +215,33 @@ def test_backend_datastore_add_delete(tgen):
             "/frr-vrf:lib/vrf",
         ]
     )
-    assert waitline(mlogp.stdout, 'now known as "mgmtd-testc"', timeout=10)
-    mlogp.kill()
+    check_backend_xpath_registry(r1, r"/frr-vrf:lib/vrf:.*mgmtd-testc.*")
 
     r1.cmd_raises('vtysh -c "conf t" -c "int foobar"')
     try:
+        #
+        # If have a failure here b/c we are now notifying at the list element
+        # level (instead of `.../state`) we have either 1) solved the multiple
+        # owner of sub-tree data problem, or 2) someone has made this change not
+        # understanding the problem (i.e., zebra deletes the list element but
+        # some other backend daemon owns state under the interface and hasn't
+        # deleted it yet).
+        #
+        # When we do have a comprehensive solution (1) for the multiple owner
+        # issue we should update this test to be more stringent and look for the
+        # list element to be deleted.
+        #
         assert waitline(
             p.stdout,
             re.escape('#OP=REPLACE: /frr-interface:lib/interface[name="foobar"]/state'),
-            timeout=10,
+            timeout=30,
         )
 
         r1.cmd_raises('vtysh -c "conf t" -c "no int foobar"')
         assert waitline(
             p.stdout,
             re.escape('#OP=DELETE: /frr-interface:lib/interface[name="foobar"]/state'),
-            timeout=10,
+            timeout=30,
         )
 
         # Now add/delete a VRF and watch for notifications
@@ -243,21 +249,115 @@ def test_backend_datastore_add_delete(tgen):
         r1.cmd_raises("ip link add red type vrf table 10")
         r1.cmd_raises('vtysh -c "conf t" -c "vrf red" -c "exit"')
 
-        wait_op_json(
-            p.stdout,
-            "REPLACE",
-            '/frr-vrf:lib/vrf[name="red"]/state',
-            '{"frr-vrf:lib":{"vrf":[{"name":"red","state":{"active":true}}]}}',
-        )
+        wait_op_json(p.stdout, "REPLACE", '/frr-vrf:lib/vrf[name="red"]')
 
         r1.cmd_raises("ip link del red")
         r1.cmd_raises('vtysh -c "conf t" -c "no vrf red"')
         wait_op_json(p.stdout, "DELETE", '/frr-vrf:lib/vrf[name="red"]')
     finally:
-        pass
         p.kill()
         r1.cmd_status('vtysh -c "conf t" -c "no vrf red"', warn=False)
         r1.cmd_status("ip link del red", warn=False)
+
+
+def test_backend_datastore_router_id(tgen):
+    if tgen.routers_have_failure():
+        pytest.skip(tgen.errors)
+
+    r1 = tgen.gears["r1"].net
+
+    check_kernel_32(r1, "11.11.11.11", 1, "")
+
+    if not have_beclient(r1):
+        pytest.skip("No mgmtd_testc")
+
+    # Start our BE client in the background
+    p = r1.popen(
+        [
+            BE_CLIENT,
+            "--timeout=30",
+            "--log=file:mgmtd-testc.log",
+            "--notify-count=0",
+            "--datastore",
+            "--listen",
+            "/frr-vrf:lib/vrf/frr-zebra:zebra/router-id",
+            "/frr-vrf:lib/vrf/frr-zebra:zebra/ipv6-router-id",
+        ]
+    )
+
+    js4_init = '{"frr-vrf:lib":{"vrf":[{"name":"default","frr-zebra:zebra":{"router-id":"1.1.1.1"}}]}}'
+    js4_chg = '{"frr-vrf:lib":{"vrf":[{"name":"default","frr-zebra:zebra":{"router-id":"1.2.3.4"}}]}}'
+    js6_new = '{"frr-vrf:lib":{"vrf":[{"name":"default","frr-zebra:zebra":{"ipv6-router-id":"aa::bb"}}]}}'
+    js6_chg = '{"frr-vrf:lib":{"vrf":[{"name":"default","frr-zebra:zebra":{"ipv6-router-id":"aa::cc"}}]}}'
+
+    try:
+        #
+        # IPv4 Router ID
+        #
+        wait_op_json(
+            p.stdout,
+            "SYNC",
+            "/frr-vrf:lib/vrf/frr-zebra:zebra/router-id",
+            js4_init,
+        )
+        r1.cmd_raises('vtysh -c "conf t" -c "router-id 1.2.3.4"')
+        wait_op_json(
+            p.stdout,
+            "REPLACE",
+            '/frr-vrf:lib/vrf[name="default"]/frr-zebra:zebra/router-id',
+            js4_chg,
+        )
+        r1.cmd_raises('vtysh -c "conf t" -c "no router-id"')
+        wait_op_json(
+            p.stdout,
+            "REPLACE",
+            '/frr-vrf:lib/vrf[name="default"]/frr-zebra:zebra/router-id',
+            js4_init,
+        )
+
+        #
+        # IPv6 Router ID
+        #
+        r1.cmd_raises('vtysh -c "conf t" -c "ipv6 router-id aa::bb"')
+        wait_op_json(
+            p.stdout,
+            "REPLACE",
+            '/frr-vrf:lib/vrf[name="default"]/frr-zebra:zebra/ipv6-router-id',
+            js6_new,
+        )
+        r1.cmd_raises('vtysh -c "conf t" -c "ipv6 router-id aa::cc"')
+        wait_op_json(
+            p.stdout,
+            "REPLACE",
+            '/frr-vrf:lib/vrf[name="default"]/frr-zebra:zebra/ipv6-router-id',
+            js6_chg,
+        )
+        # r1.cmd_raises('vtysh -c "conf t" -c "no ipv6 router-id"')
+        # wait_op_json(
+        #     p.stdout,
+        #     "DELETE",
+        #     '/frr-vrf:lib/vrf[name="default"]/frr-zebra:zebra/ipv6-router-id',
+        # )
+    finally:
+        p.kill()
+        r1.cmd_status('vtysh -c "conf t" -c "no ipv6 router-id"', warn=False)
+        r1.cmd_status('vtysh -c "conf t" -c "no router-id"', warn=False)
+
+
+@retry(retry_timeout=30, initial_wait=0.5)
+def check_filters(r1, selectors):
+    showcmd = 'vtysh -c "show mgmt get-data /frr-backend:clients/client/state/notify-selectors"'
+    jsout = json.loads(r1.cmd_raises(showcmd))
+
+    if not selectors:
+        expected = {}
+        result = json_cmp(jsout, expected, exact=True)
+        assert result is None
+        return
+    state = {"notify-selectors": selectors}
+    expected = {"frr-backend:clients": {"client": [{"name": "zebra", "state": state}]}}
+    result = json_cmp(jsout, expected, exact=True)
+    assert result is None
 
 
 def test_datastore_backend_filters(tgen):
@@ -266,33 +366,59 @@ def test_datastore_backend_filters(tgen):
 
     r1 = tgen.gears["r1"].net
 
+    if not have_beclient(r1):
+        pytest.skip("No mgmtd_testc")
+
     check_kernel_32(r1, "11.11.11.11", 1, "")
 
-    # Start our FE client in the background
-    p = r1.popen(
-        [FE_CLIENT, "--datastore", "--listen=/frr-interface:lib/interface/state"]
-    )
-    assert waitline(p.stderr, "Connected", timeout=10)
-    time.sleep(1)
-
+    p = None
+    p2 = None
     try:
-        output = r1.cmd_raises(
-            'vtysh -c "show mgmt get-data /frr-backend:clients/client/state/notify-selectors"'
+        step("Verify no backend notify selectors")
+        check_filters(r1, [])
+
+        step("Start FE client and verify notify selector present in zebra")
+        p = r1.popen(
+            [
+                FE_CLIENT,
+                "--notify-count=0",
+                "--datastore",
+                "--listen=/frr-interface:lib/interface/state",
+            ]
         )
-        jsout = json.loads(output)
+        check_filters(r1, ["/frr-interface:lib/interface/state"])
 
-        #
-        # Verify only zebra has the notify selector as it's the only provider currently
-        #
-        state = {"notify-selectors": ["/frr-interface:lib/interface/state"]}
-        expected = {
-            "frr-backend:clients": {"client": [{"name": "zebra", "state": state}]}
-        }
+        step("Start BE client and verify both notify selectors present in zebra")
+        p2 = r1.popen(
+            [
+                BE_CLIENT,
+                "--log=file:mgmtd-testc.log",
+                "--notify-count=0",
+                "--datastore",
+                "--listen",
+                "/frr-interface:lib/interface",
+            ]
+        )
+        selectors = [
+            "/frr-interface:lib/interface",
+            "/frr-interface:lib/interface/state",
+        ]
+        check_filters(r1, selectors)
 
-        result = json_cmp(jsout, expected, exact=True)
-        assert result is None
+        step("Kill BE client and verify only FE notify selector present in zebra")
+        p2.kill()
+        p2 = None
+        check_filters(r1, ["/frr-interface:lib/interface/state"])
+
+        step("Kill FE client and verify no notify selectors present in zebra")
+        p.kill()
+        p = None
+        check_filters(r1, [])
     except Exception as error:
         logging.error("got exception: %s", error)
         raise
     finally:
-        p.kill()
+        if p2:
+            p2.kill()
+        if p:
+            p.kill()
